XMLRPCServer: make XMLRPCServer able to listen on IPv6 addresses
SimpleXMLRPCServer is derived from SocketServer.TCPServer of which the
default address_family is socket.AF_INET, that makes it could only
listen on IPv4 connections. This patch wraps it and extends its
ability on IPv6 networks.

How to test:

  1. Build and install VDSM RPMs
  2. Edit /etc/vdsm/vdsm.conf, change "addresses/management_ip = ::",
and "vars/jsonrpc_enable = false"(as known to break JSONRPC)
  3. Restart VDSM daemon
  4. "netstat -nlp6t" or "ss -nlp6t" should indicate vdsm XMLRPC server
is listening on "::"
  5. "vdsClient -s '[::1]:54321' getVdsCaps" succeeds

Change-Id: I50ab0ba59ac696da2d992afe453c6d3ceb111e94
Signed-off-by: Petr Sebek <psebek@redhat.com>
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Signed-off-by: huntxu <mhuntxu@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/11520
Reviewed-by: Dan Kenigsberg <danken@redhat.com>