Fix bridgeless net regression
The commit:
Ie8bb5fd9fc6ad69cf120d62264d73490b65b82d7
http://gerrit.ovirt.org/#/c/11356/4/

That commit introduced a bug that prevents the user from configuring
a bridgeless network with dhcp. That is so because after the change,
if ipaddr is None, which is the most usual case when configuring a
network for dhcp, bootproto will be overriden by what was on the
original config file (quite probably None).

Change-Id: I2881791f7b2b23cdcea25b73d3eb44de2d2addc8
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16086
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>