vdscli: update cannonize helper function for IPv6 environment
The original approach considers any string contains a colon to have
a port information which is not true in IPv6 environment.

The new approach assumes the input string follows the format
'host [ ":" port ]' in rfc3986 as it would be then used in a URI, gives
a default port if the input doesn't supply one. This means IPv6
addresses should be enclosed within square brackets ("[" and "]").

How to test:
* The following commands should all work correctly in a default setup
  1) for IPv4 environment:
    * vdsClient -s 127.0.0.1 getVdsCaps
    * vdsClient -s 127.0.0.1:54321 getVdsCaps
    * vdsClient -s localhost getVdsCaps
    * vdsClient -s localhost:54321 getVdsCaps
  2) for IPv6 environment:
    * vdsClient -s [::1] getVdsCaps
    * vdsClient -s [::1]:54321 getVdsCaps
    * vdsClient -s localhost6 getVdsCaps
    * vdsClient -s localhost6:54321 getVdsCaps
    * vdsClient -s ['IPv6 link-local addr'%ovirtmgmt] getVdsCaps
    * vdsClient -s ['IPv6 link-local addr'%ovirtmgmt]:54321 getVdsCaps

Change-Id: I7c450aaa141ed2877a2d380f17959d2dd8fa77c9
Signed-off-by: huntxu <mhuntxu@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/16225
Reviewed-by: Petr Šebek <psebek@redhat.com>
Tested-by: Petr Šebek <psebek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>