API: Add free memory to stats
currently, only memAvailable is exposed by API. But this api is a
calculated value and so it may return negative value in case ksm
is up and pages are shred among VMs.

The new "memFree" returns the true free memory on the host.
i.e. 'MemFree' + 'Cached' + 'Buffers' (from meminfo)

Change-Id: I0e5ebf8b40b728424ba92bf8131fff51b45893bd
Bug-Url: https://bugzilla.redhat.com/977758
Signed-off-by: Ofri Masad <omasad@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16253
Reviewed-by: Doron Fediuck <dfediuck@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>