vdsm-tool: service management work with previous systemd release
The service management assumes systemctl automatically adds ".service"
suffix to the service name. This is true for the latest SystemD, but
before systemd-44-20.fc17 [1], this feature is not available.

This patch adds ".service" suffix before sending the service name to
systemctl, so it does not depend on the auto-suffixing feature any more.

[1] https://lists.fedoraproject.org/pipermail/package-announce/2012-October/090358.html

Change-Id: Ice83296bd88ed5bb13630694560cde58b050edbb
Reported-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/16379
Reviewed-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>