before/after_update_device hook now receives custom properties
Patch Set 2:
Patch set 1 took the custom properties from the device. But
that means that the custom properties past in the command
are ignored. Patch set 2 fixes this: The custom properties
are now always taken from the verb's arguments, and the
device's custom properties are updated.

Patch Set 3:
Moved the update to the in-memory cache of the device's
custom properties to a more sensible location.

Change-Id: Ife0f617fe7e40104de3dc5bcad28b2de04f1d446
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16408
Reviewed-by: Dan Kenigsberg <danken@redhat.com>