Multiple Gateways[3/2]: Re-did comms between dhclient and vdsm
SElinux disallows dhclient to change routing table information directly,
and so we needed a different way for dhclient to communicate with vdsm.
We did it via files and inotify.
dhclient writes to /var/run/vdsm/sourceRoutes with all of the info
of the DHCP response. supervdsm checks that folder upon startup
(in case dhclient operated while supervdsm was down) and registers
an inotify event handler on that folder on a separate thread.

Patch Set 3:
* Re-added libvirt check when configuring a network
* Added sourceRouteThread.py, /var/run/vdsm/sourceRoutes dir to
  spec and make files
* Capitalized DHClientEventHandler, removed os.path.join usage
* Files created in /var/run/vdsm/sourceRoutes are now named $timeStamp
  and we now read the folder in alphabetical order during vdsm start up
  so that we handle dhcp responses by their received order

Patch Set 4:
* Removed argparse from spec file
* Changed from file created to file closed to fix possible race
* Answered Giuseppe's and Dan's reviews

Patch Set 5:
* Fixed pep8 issue
* Fixed bug with signature of the event handler. It needed two
  versions: One I call manually, and one inotify calls

Change-Id: Ib7a47a680bfbfe285f708290de97e8f08714695a
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16461
Reviewed-by: Dan Kenigsberg <danken@redhat.com>