Multiple Gateways[4/2]: Fixed ifdown bug for DHCP interfaces
If the admin manually ifdown'd an interface of ours following
patch 3 in the series, we failed to remove the routes and rules
we created during ifup. The reason is that since patch 3 we
are working with files. The dhclient hook created a file during
ifdown and finished the ifdown operation before VDSM got the
notification that the file was created. By the time VDSM got
the notification the interface was already down and thus could
not find any information about the interface.

This patch fixes that bug by adding the device to the rule
while adding the rule, so that we may use the device name
to look up all rules created for that device.

Patch Set 2:
* Renamed Rule.inDevice to Rule.srcDevice
* Removed sourceRoute.py:_getNetwork

Change-Id: If1ed9efc548ec0342db9ea0ae4692d2c0aec93df
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16552
Reviewed-by: Dan Kenigsberg <danken@redhat.com>