spec: restart systemd-modules-load.service on post installation
It's observed that the needed modules are not loaded after vdsm is
installed without rebooting system. That's because we reply on
systemd-modules-load.service to load the modules required by vdsm,
but systemd-modules-load.service can't detect the new config
file added by vdsm. It should not be a problem if the vdsm host
is added to ovirt engine. So this change only make it more friendly
for people who run vdsm in standalone mode.

Change-Id: I7786a6d73307369f72ea83f27c3a128ba9d99159
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/16610
Reviewed-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>