Add bootproto during delNetwork
bootproto isn't sent from the engine for the delNetwork command.
However, the multiple gateways feature code needs to behave
differently during delNetwork as a function of the bootproto.

This patch gets the bootproto from the cfg file. When we
add other configurators we'll have to deal with this.

Patch Set 2:
* Moved getBootProto from configNetwork to netinfo

Change-Id: Ibe7c39ed1eeb94e806b168bd346dbcd51abd8ffa
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16678
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>