vdsm: /usr/lib/tmpfiles.d/vdsm.conf
tmpfiles.d is a service in Fedora 15 and later for managing temporary files and
runtime directories for daemons. Directories /var/run and /var/lock are symlinks
into the /run tmpfs filesystem. As such, they are created empty on every reboot.
For files intended to be placed into those directories, this should normally
not pose any problems. For directories, however, we often need to create the
directories ahead of time. This is best done using the tmpfiles.d mechanism.

Change-Id: Iea9ec4b7ba1e362c1141c4e0cb734d2b1520dbfd
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/15707
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>