spec: fix that vdsmd.service is not enabled after installation
For fedora18+, we use the helper macro to replace the manual systemd
scriptlets. But for post installation, %systemd_post is expanded as
below:
	/usr/bin/systemctl preset vdsmd.service >/dev/null 2>&1 || :
So vdsm service will be disabled after installation because the default
systemd preset setting disable all services which are not explictly
enabled in the configuration files.

We can ship a preset configuration file to enable vdsmd service, but
reverting to manual systemd scriptlets is simpler and doesn't lose any
benefit.

Change-Id: I4ba726296d510c1281bc9425e9f420b911ec45f7
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/16754
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>