Unified network persistence [1/4] - Save running config
This patch introduces the new persistence model for vdsm networking.
It is meant to provide a single reliable way abstracting persistence
out of the netconf configurators as much as possible.

To achieve its purpose, it stores the network actions as setupNetwork
parameters serialized in json which are then used for rollback and
initialization.

The first patch in the series saves the current state of networks,
nics and bonds to a folder.
The second patch copies the contents of the running config folder
to a startup config folder
The third patch uses the startup config folder to restore network
configuration to the host upon host boot up.

Change-Id: I7137a96f84abd2c5e532c6c37737e36ef17567a9
Signed-off-by: Assaf Muller <amuller@redhat.com>
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/16699
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>