Unified network persistence [2/4] Respond to setSafeNetworkConfig
This patch adds setSafeNetwork capabilities to the unified
persistence model. At this stage, it is able to store the running
configuration to /var/lib/vdsm/persistence/netconf/{nets,bonds}
in an atomic way so that it will be available after reboot.

In the following patches there should be added a service that
calls a configurator flushing system that deletes the configurator
persistence bits before networking goes up and instead uses the
unified persistence to restore the network configuration.

Change-Id: I320677e40ff5b11da684d3ab7195d018135356b2
Signed-off-by: Assaf Muller <amuller@redhat.com>
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17009
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>