Source routes files now deleted on non-VDSM interfaces
Files in /var/run/vdsm/sourceRoutes are created by a VDSM DHCP
hook for all interfaces that release or get a DHCP configuration.
sourceRouteThread.py configures source routing on those interfaces.
However, first it checks that the interface in question is known
to VDSM. Previously the files weren't deleted in such a case.

Change-Id: I7bb4f4859cc6a8e8427d655de2a54fdc61e67430
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17174
Reviewed-by: Dan Kenigsberg <danken@redhat.com>