fix setupNets 'remove' of non-existent net
setupNetworks was assuming that if a net was not in vdsm nets nor in
libvirt nets it was a net that we were trying to add when, in fact,
it could very well be a non-existent network that was incorrectly
marked for removal.

The fix checks the case specified above and raises the same
configNetworkError that delNetwork does on similar circumstances.

Change-Id: If79c46cbc21c5bb451d69c2acd73f6e1ecf330bf
Bug-url: https://bugzilla.redhat.com/977843
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17254
Reviewed-by: Assaf Muller <amuller@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>