do not rely on default values on restoreFlags invocation
The signature of libvirt's verb 'restoreFlags' changed - in previous
versions it didn't set default values for its optional arguments. this
patch changes how we invoke this verb to set the 'flags' argument to 0
explicitly in order to be backward compatible with libvirt versions that
don't set default values for 'restoreFlags' arguments.

Change-Id: I0dd348370c5dd6adfad813c7f01904c170f523bf
Bug-Url: https://bugzilla.redhat.com/988229
Signed-off-by: Arik Hadas <ahadas@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17298
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>