Don't down bonds and nics when adding vlan or resizing
When adding vlans on top of devices, a configure of the underlying
device incurs in ifdown and ifup, which in case of that device
backing other vlans, will cause a network interruption.

This patch makes that nics are only reconfigured (if a vlan is being
put on top of them) when the interface was already down. This works
because any extra config that would need to be applied for the new
action needs only application on the vlan device itself.

For Bonds, we only need to reconfigure in case the bonding options
differ or the slaves don't match, in any other case, it's the same
as with nics.

Change-Id: Id3075a2e65f06416c840c6a98689b77555b22e5d
Bug-Url: https://bugzilla.redhat.com/982632
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17491
Reviewed-by: Dan Kenigsberg <danken@redhat.com>