SecureXMLRPCServer: Work with Official M2Crypto Release
VDSM uses the "settimeout" method of M2Crypto.SSL.Connection. This
method is defined by a patch shipped in Fedora and RHEL m2crypto RPM.
The "settimeout" patch is not available in official code, so other
distro such as Ubuntu does not provide "settimeout".

We can detect if the "Connection" class provides "settimeout" and work
with both RPM and official release of M2Crypto. If it provides
"settimeout", use it, otherwise falls back to use "settimeout" from the
socket object.

Change-Id: Id8db9140f27ca17adf4db0275188511257ba9116
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/17809
Reviewed-by: Dan Kenigsberg <danken@redhat.com>