vdsm: Always retrieve alias for Balloon and Console devices
We haven't been retrieving the alias value for console devices which
now starts to bite us due to the assumption that all VmDevice objects,
have to have an alias attribute during the migration.

Due to a bug in the handling for Balloon devices this also happened.

This commit fixes the problem for the future by retrieving the aliases.

Change-Id: Ic828d13761b26b3ec01b604f431050a611bce01c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=988065
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17602
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17686