Handling signals interrupts, using pidofproc and prevent lock over systemd
Under start lock, if signal is received, we should clean the lock file and
stop the instance. During stop we should not stop the process until
respawn and vdsm instances are cleared.

Instead of checking the existence of respawn.pid and vdsmd.pid files, using
pidofproc checks if the processes are actually run. The deletion of the
respawn.pid file is redundant and leads to conflicts.

Over systemd we don't require to use locking at all. Systemd takes care of
that.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=994912
Change-Id: I9b58672957840c9dc5d13dce6af8de1717a9cb2d
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/17926
Tested-by: Elad Ben Aharon <eladba1990@gmail.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>