Functional Tests [2/3]: Added ipwrapper.ruleExists+routeExists
Added functional tests for both functions as well.

We can now parse rules that contain '[detached]' - If you add
rules for devices that don't exist (Or add rules for devices
that do exist and then delete the device) then 'ip rule' outputs
the rule with '[detached]'. We couldn't parse such rules.
This is important because when cleaning up rules that the test
added we first get rid of the dummy nic and only then get
rid of rules.

Change-Id: Iaf7c28e85f53f51879750308f3a9f93fc635cf33
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18040
Reviewed-by: Dan Kenigsberg <danken@redhat.com>