gluster: bug fix for listing and adding hook
Earlier, with regard to gluster hooks there would be definitely a pre and post
directory for each and every gluster command. But presently, if either pre
or post script is not present there will be no pre or post directories also.
Therefore, if there is no such directories available the gluster hook list
will simply omit it.
When a gluster hook is added if pre or post directories are absent, they
will be created accordingly, then the hook will be added.

Change-Id: I470470da153edef3fe7eb3a7d41a775bac208449
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=999795
Signed-off-by: Timothy Asir <tjeyasin@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18345
Reviewed-by: Dan Kenigsberg <danken@redhat.com>