Get rid of mutables(lists) as default parameters
In Python having a mutable as a default method parameter is
dangerous due to the fact that the mutables are created on
module creation rather than on method invokation. Thus, a
call to a method could pollute the environment for the next
call.

This patch cleans most instances of that happening.

Change-Id: I911485cf30d587f5b6b8bf44cf3552c1517abfd5
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18356
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Petr Šebek <psebek@redhat.com>
Tested-by: Petr Šebek <psebek@redhat.com>