vm: add the transient disk support
A transient disk is a temporary qcow layer based on a (usually
internal) volume that is attached to a VM in read-write mode.
All changes stored in the transient disk will be lost as soon
as the disk is hot-unplugged or when the VM exits.
Its primary use is to access in read-write mode a snapshot
(read-only volume).
Libvirt doesn't support transient disks yet, hence this patch
includes also all the required parts to fill that gap.

Change-Id: I3dfec35e324c47d8c86a965947e3ae4ae48c7524
Signed-off-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18326
Reviewed-by: Dan Kenigsberg <danken@redhat.com>