fix getVlanID returning an int issues
The fact that getVlanId now returns an int causes issues across
the networking codebase due to the fact that there was a lot of
concatenation of the kind:

    vlanDev = ifacename + '.' + vlanId

which of course causes a TypeError due to trying to concatenate
string and int.

Additionally, there was code like:

if (not) vlan:

which now that vlan is int, will be evaluated as False when we
use vlan with id 0.

Finally, this patch makes all internal dealings with vlan ids
to be converted to integer on reception.

Change-Id: I0221bd6ba573c564052bac468d119c2aaceef6e0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18449
Reviewed-by: Amador Pahim <apahim@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>