Remove bonds before editing and adding
The current implementation of handleBondings relies on the
non-deterministic ordering of the bondings dictionary for choosing
which bonds should be handled first.

The correct ordering is to first delete all the bonds to remove,
then edit and finally add new bonds. This ensures that slaves
of bonds to remove are properly freed before being added to existing
or new bonds.

Change-Id: I3ea081fe7760b52f6febb4959f69851b2cf4c2a1
Bug-Url: https://bugzilla.redhat.com/987525
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18500
Reviewed-by: Assaf Muller <amuller@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>