fix testDelNetworkWithMTU
If a network with an MTU lesser than the default is created on top of an
existing nic or bond, the new MTU is not set on the underlying device.
This makes sense if the high MTU of the device is being used by
something else, but not if there are no users to the device.

There is a caveat to this fix, though. If someone outside of Vdsm has
been using the underlying device for other purposes (such as another
vlan with a higher MTU), Vdsm would now change the MTU while it takes
ownership on the device. This should be of no surprise, since when Vdsm
owns a device, it truly owns it, and may control all of its attributes.

Change-Id: I151447250159df389a159d3823a9bd41cb51c211
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18714
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Giuseppe Vallarelli <gvallare@redhat.com>