Removing automatic restart of services in libvirt-reconfigure
libvirt-reconfigure also restarted libvirtd as post reconfigure operation. The
restart is required but there is no reason and its not expected to perform it
automatically as part of the reconfigure operation.

Restart also required to supervdsmd service (as described in the
Bug-Url). Instead of restarting both, this patch prints message to user to
preform the restart manually.
To make it easier and centralized the patch also adds
libvirt-configure-services-restart command that restart related services
after reconfigure.

For backward compatibility with systemd, this also update
systemd-vdsmd script to restart libvirtd and supervdsmd services in
previous versions.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=999853
Change-Id: If3711da62d6c196978fb4f5e89f1bce4b7500470
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18780
Reviewed-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>