Modification in lockfile usage mechanism as part of sysv init script
The problem:
Start can run if lock is free and respawn pid is not exist. Currently we keep
holding the lock until the service is stopped. It means, that if the
process that was initiated (the respawn prcoess) was crashed, we still hold
the lock (as we didn't stop it by service stop) and this blocks us from
initiating new instance. We acquire lock when start, and release it only after
stop.

The solution:
Using flock to avoid multi instances of the init script. Vdsm execution
is performed in subprocess and it releases the locked fd before starting
the respawn process.

Change-Id: Ic74c9cf9c2c520e94a98b99fdb970b544b780ba3
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18875
Reviewed-by: Dan Kenigsberg <danken@redhat.com>