Packaging: libvirt plugin of selinux and sanlock usage is configurable at build time
VDSM makes use of these two plugins. However on platform such as Ubuntu,
libvirt is compiled without selinux and sanlock plugin. In this patch we
make these plugins configurable at build time by

--enable-libvirt-selinux
--enable-libvirt-sanlock

By default they are both enabled.

Because of non standard use of substitution (not using AC_OUTPUT) the
scripts that are effected will not be regenerated unless make clean is
called, out of the scope of this patch.

Change-Id: I0953942999e5e77054c3c8f4f2b48e9cfd31fa61
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/18921
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>