Packaging: Remove hard coded 12- prefix of vdsm-lvm.rules from upstream
vdsm-lvm.rules.in is to fix lvm block device permissions, it should be
applied after the system default lvm rules. On RH family distributions,
12 is correct because system default is 11-dm-lvm.rules. On Ubuntu, 12 is
incorrect and we use 61-vdsm.lvm.rules. This patch removes the 12-prefix
from the upstream file name, and leaves the downstream packaging script
to decide the correct installed name.

Change-Id: I65ae0fa4d83bba0e1622f70d5c8bee0e2d8742dd
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/19006
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>