Packaging: Fix help string and usage for --disable-libvirt-sanlock
Help strings in configure.ac for --disable-libvirt-sanlock and
--disable-libvirt-selinux are as follow.

  --disable-libvirt-sanlock
                          disable libvirt sanlock plugin [default=yes]
  --disable-libvirt-selinux
                          disable libvirt selinux plugin [default=yes]

From simplicity and disambiguation, we remove "[default=yes]". This
patch changes the help string to as follow

  --disable-libvirt-sanlock
                          disable libvirt sanlock plugin
  --disable-libvirt-selinux
                          disable libvirt selinux plugin

So there is no ambiguity of whether the feature is disbale or enabled
by default. We also avoid suggesting user to use
"--disable-libvirt-sanlock=yes|no" which is a bit confusing (at least to
me). The user can just use "--disable-libvirt-sanlock" to disable it.

This patch also fix the code logic related to ENABLE_LIBVIRT_SELINUX in
libvirt_configure.sh.in to make it agree with the help string.

Change-Id: I935693a8b8e7c371c653853b9524c9a774bca6fd
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/19005
Reviewed-by: Dan Kenigsberg <danken@redhat.com>