sp: update lease parameters on masterMigrate
Given the current API the lease parameters should be considered pool
metadata (since they're passed to vdsm in createStoragePool and in
reconstructMaster). On the contrary the implementation is (rightfully,
you might argue) storing these values in the domain metadata.

On master migrate we should copy the lease parameters from the old
master to the new one in order to maintain the appropriate values.

In this patch:
- update lease parameters on masterMigrate
- relax the default parameters

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=999343
Change-Id: I0e5935e8a7081c391702f63d450621650769bb92
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19176
Reviewed-by: Dan Kenigsberg <danken@redhat.com>