caps: coarse-grain libvirt.rpm may be missing
As of http://gerrit.ovirt.org/15761 we no longer require the legacy
libvirt.rpm to be installed on Fedora, so getCaps should not look for
it. Instead, it should report the information of the installed
libvirt-daemon-kvm.rpm.

Without this patch we get an annoying
    (_getKeyPackages) rpm package libvirt not found
log message, and do not report libvirt's version.

Change-Id: I733a1a2411c910caa030e435c2a7f751a8e4dad4
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19295
Reviewed-by: mooli tayer <mtayer@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21363
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>