Log client address on SSL errors
We used to write to the log the IP address of the client when an SSL
error occurred. With the introduction of cached SSL sessions this
stopped working. This patch modifies the SSLServerSocket class that was
introduced in order to implement cached SSL sessiosn so that when an SSL
error is detected while accepting a connection the IP address of the
client will be added to the message of the exception, and thus written
to the log as part of the traceback, for example:

  BindingXMLRPC::ERROR::2013-09-17
  20:34:25,901::BindingXMLRPC::72::vds::(threaded_start) xml-rpc handler exception
  Traceback (most recent call last):
    File "/usr/share/vdsm/BindingXMLRPC.py", line 68, in threaded_start
      self.server.handle_request()
    File "/usr/lib64/python2.7/SocketServer.py", line 280, in handle_request
      self._handle_request_noblock()
    File "/usr/lib64/python2.7/SocketServer.py", line 290, in _handle_request_noblock
      request, client_address = self.get_request()
    File "/usr/lib64/python2.7/SocketServer.py", line 463, in get_request
      return self.socket.accept()
    File "/usr/lib64/python2.7/site-packages/vdsm/SecureXMLRPCServer.py", line 128, in accept
      raise SSL.SSLError("%s, client %s" % (e, address[0]))
  SSLError: sslv3 alert certificate unknown, client 192.168.122.1

Note the text "client 192.168.122.1" added to the last line of the
traceback.

Change-Id: Icaed8689bc8bb5bee5fed450646b2c5e9ef42bc1
Bug-Url: https://bugzilla.redhat.com/983630
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19332
Reviewed-by: Dan Kenigsberg <danken@redhat.com>