iscsiadm: Filter out IPv6 addresses when discovering targets
This patch is not diminishing any capabilities we have, since we
currently don't support ipV6 targets in vdsm.

When discovering targets on Netapp storage we get a return value that
includes both IPv4 and IPv6 addresses.

The only thing I've changed is preventing vdsm from failing discovery
because of getting an unexpected response from iscsiadm (our parser is
very hardcoded).

If we decide to support ipV6 addresses for discovery, login and managing
the connection then it needs to be addressed separately.

Change-Id: I2f6d60e780e9ac50041cdcaa0c45394e392b52f4
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19431
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>