net tests: setupNetworks over exiting USED bond
Currently testSetupNetworksAddOverExistingBond fails, but only because
it is too strict: it does not allow the bonding device to change its
state even when it has no users. However, if the bond is not used, we
actually WANT it to be reconfigured (e.g. if its MTU is to be set to a
lower value).

Thus the tests is amended so that it would fail only if bond0 changes
its operstate while it is being used by a birdgeless network.

Change-Id: Ibc5159665a23863cc997390cfc6dc9100a33eca2
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19573
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>