Workaround to avoid password request when starting vdsmd service
During start of vdsmd we shutdown conflicting services. As part of this
we try to stop libvirt-guests. `service libvirt-guests stop` tries to
access libvirtd which requires sasl authentication.
This patch removes libvirt-guests lockfile instead of stopping the service.

The patch should be reverted after a fix for:
https://bugzilla.redhat.com/show_bug.cgi?id=1011981 will be available
(Comment was added above the change to keep tracking the issue)

The workaround was part of vdsmd init script already in the past and was
removed by mistake.

Change-Id: If7a3ad4db377f57a9d7fd34afc33d1add25c967f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1010693
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19644
Reviewed-by: Dan Kenigsberg <danken@redhat.com>