Manage and fix test dummies handling -> module level
We can handle the dummies at the global level so that less work needs
to be done per test and the cleanupNetwork decorator doesn't miss
the dummies that were previously already destroyed at that stage.

The reutilization of the dummy interfaces exposed two bugs in the
current tests in their cleaning up phase that this patch addresses
as well.

Change-Id: I5c040fc3175b5760728d749578728be60c75cb9b
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19650
Reviewed-by: Dan Kenigsberg <danken@redhat.com>