fix premature selinux dependency
Vdsm v4.13.0 places its storage repository under /var/run/vdsm/storage
instead of the non-standard trademarked /rhev/data-center. This change
requires a compatible change to selinux which would be available only
when selinux-policy-3.7.19-195.el6.13 is released.

Until then, we try to hack the issue on EL6 by tweaking the policy on vdsm installation with

    semanage fcontext -a -s system_u -t virt_var_run_t -r s0 '/var/run/vdsm(/.*)?'

This patch should be promptly reverted when
selinux-policy-3.7.19-195.el6.13 is released.

Bug-Url: https://bugzilla.redhat.com/1021147
Change-Id: Iea625d7c39035055b246357030372430b4bddf68
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19807
Reviewed-by: Eyal Edri <eedri@redhat.com>
Reviewed-by: Ohad Basan <obasan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20313
Tested-by: Fabian Deutsch <fabiand@fedoraproject.org>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>