Removing libvirt_configure from pre-start and ask for manual run
Before starting vdsmd we need to add vdsm configuration to
libvirtd.conf. This part is performed by vdsm-tool libvirt_configure
verb. Currently we call it as part of init script pre-start function,
but doing that must follow restart of depended services as libvirtd. We
don't want to restart services as part of vdsmd init script, so this
patch detach the auto reconfigure and report to user to perform it
manually if attempt to start vdsmd without having libvirtd.conf
configured.
Performing start\stop actions on external services during service start
is invalid and can lead to bugs in different service managers.

Change-Id: Icf70a749454ea341d5b52220e16e9567b90431a0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1013371
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19737
Reviewed-by: Dan Kenigsberg <danken@redhat.com>