Bump kernel requires version to pass testSetupNetworksMtus on F19
It's found that testSetupNetworksMtus fails on F19 because of the
broken promiscuity reference counting issue in kernel. It has been
fixed in kernel-3.11.3-201.fc19. With the updated kernel, the test
testSetupNetworksMtus can pass. For detailed information, please see:
https://bugzilla.redhat.com/show_bug.cgi?id=1005567

This problem doesn't exist in RHEL6 because it's triggered by a new
kernel change. So we don't need any fix for rhel kernel.

Change-Id: I161b34b0a67c88fe18c21e9857841639f7001098
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/19974
Reviewed-by: Dan Kenigsberg <danken@redhat.com>